# Further optimize the box rearrange create algorithm