Optimize the box rearrange create algorithm by reducing
the number of integer divisions used to convert a 1-D
index into a coordinate value.

For high-res E3SM cases using short PIO strides, this
optimization can greatly reduce the run time when BOX
rearranger is being used to read or write large-size
variables.

Also see PR #112 that optimized the same algorithm.

Fixes #179