@dqwu 2019-03-29
[High-res G case run with 9600 MPI tasks, BOX rearranger, PIO stride 1]
Without this optimization
"a_i:PIO:box_rearrange_create" 3.465 sec
"r_i:PIO:box_rearrange_create" 0.849 sec
"o_i:PIO:box_rearrange_create" 260.449 sec
"i_i:PIO:box_rearrange_create" 62.801 sec
"o:PIO:box_rearrange_create"   355.447 sec
Total: 683.011 sec

With this optimization
"a_i:PIO:box_rearrange_create" 3.459 sec
"r_i:PIO:box_rearrange_create" 0.844 sec
"o_i:PIO:box_rearrange_create" 43.200 sec
"i_i:PIO:box_rearrange_create" 22.776 sec
"o:PIO:box_rearrange_create"   42.637 sec
Total: 112.916 sec

[High-res G case run with 9600 MPI tasks, BOX rearranger, PIO stride 64]
Without this optimization
"a_i:PIO:box_rearrange_create" 1.168 sec
"r_i:PIO:box_rearrange_create" 0.254 sec
"o_i:PIO:box_rearrange_create" 13.089 sec
"i_i:PIO:box_rearrange_create" 7.187 sec
"o:PIO:box_rearrange_create"   8.594 sec
Total: 30.292 sec

With this optimization
"a_i:PIO:box_rearrange_create" 1.155 sec
"r_i:PIO:box_rearrange_create" 0.243 sec
"o_i:PIO:box_rearrange_create" 11.697 sec
"i_i:PIO:box_rearrange_create" 6.539 sec
"o:PIO:box_rearrange_create"   4.005 sec
Total: 23.639 sec