@dqwu 2019-03-29
[High-res F case run with 21600 MPI tasks, BOX rearranger, PIO stride 1]
Largest dimension space: 72 x 777602 = 55,987,344

Without this optimization
"a_i:PIO:box_rearrange_create" 47.548 sec
"l_i:PIO:box_rearrange_create" 80.355 sec
"o_i:PIO:box_rearrange_create" 8.134 sec
"i_i:PIO:box_rearrange_create" 7.929 sec
"a:PIO:box_rearrange_create"   62.799 sec
"l:PIO:box_rearrange_create"   8.639 sec
Total: 215.404 sec

With this optimization
"a_i:PIO:box_rearrange_create" 38.698 sec
"l_i:PIO:box_rearrange_create" 78.939 sec
"o_i:PIO:box_rearrange_create" 8.092 sec
"i_i:PIO:box_rearrange_create" 7.968 sec
"a:PIO:box_rearrange_create"   55.795 sec
"l:PIO:box_rearrange_create"   8.290 sec
Total: 197.782 sec

[High-res F case run with 21600 MPI tasks, BOX rearranger, PIO stride 64]
Largest dimension space: 72 x 777602 = 55,987,344

Without this optimization
"a_i:PIO:box_rearrange_create" 5.066 sec
"l_i:PIO:box_rearrange_create" 8.626 sec
"o_i:PIO:box_rearrange_create" 0.673 sec
"i_i:PIO:box_rearrange_create" 0.546 sec
"a:PIO:box_rearrange_create"   4.397 sec
"l:PIO:box_rearrange_create"   1.084 sec
Total: 20.392 sec

With this optimization
"a_i:PIO:box_rearrange_create" 4.502 sec
"l_i:PIO:box_rearrange_create" 7.764 sec
"o_i:PIO:box_rearrange_create" 0.658 sec
"i_i:PIO:box_rearrange_create" 0.567 sec
"a:PIO:box_rearrange_create"   4.383 sec
"l:PIO:box_rearrange_create"   1.066 sec
Total: 18.94 sec