@dqwu 2019-03-29
@jayeshkrishna 
PR rebased with your suggested changes, I will rerun high-res F and G cases on Cori.

Note, idx_to_dim_list() and coord_to_lindex() are not marked as "static", since they are also called in tests/cunit/test_rearr.c