@dqwu 2019-04-01
@jayeshkrishna 
With the current changes (inline function call instead of explicitly inlining),
total box_rearrange_create time does not change much:
For PIO stride 1, about 2% to 4% increase
For PIO stride 64, about 12% to 14% decrease

[High-res G case run with 9600 MPI tasks, BOX rearranger, PIO stride 1]
"a_i:PIO:box_rearrange_create" 3.688 sec
"r_i:PIO:box_rearrange_create" 0.822 sec
"o_i:PIO:box_rearrange_create" 45.875 sec
"i_i:PIO:box_rearrange_create" 23.251 sec
"o:PIO:box_rearrange_create" 44.041 sec
Total: 117.677 sec (was 112.916 sec, +4.2%)

[High-res G case run with 9600 MPI tasks, BOX rearranger, PIO stride 64]
"a_i:PIO:box_rearrange_create" 1.049 sec
"r_i:PIO:box_rearrange_create" 0.220 sec
"o_i:PIO:box_rearrange_create" 9.330 sec
"i_i:PIO:box_rearrange_create" 5.779 sec
"o:PIO:box_rearrange_create" 3.934 sec
Total: 20.312 sec (was 23.639 sec, -14.1%)

[High-res F case run with 21600 MPI tasks, BOX rearranger, PIO stride 1]
"a_i:PIO:box_rearrange_create" 38.303 sec
"l_i:PIO:box_rearrange_create" 83.260 sec
"o_i:PIO:box_rearrange_create" 8.130 sec
"i_i:PIO:box_rearrange_create" 8.416 sec
"a:PIO:box_rearrange_create" 55.274 sec
"l:PIO:box_rearrange_create" 8.290 sec
Total: 201.673 sec (was 197.782 sec, +2.0%)

[High-res F case run with 21600 MPI tasks, BOX rearranger, PIO stride 64]
"a_i:PIO:box_rearrange_create" 3.971 sec
"l_i:PIO:box_rearrange_create" 7.168 sec
"o_i:PIO:box_rearrange_create" 0.509 sec
"i_i:PIO:box_rearrange_create" 0.472 sec
"a:PIO:box_rearrange_create" 3.641 sec
"l:PIO:box_rearrange_create" 0.924 sec
Total: 16.685 sec (was 18.94 sec, -12.0%)