@jayeshkrishna 2019-04-02
Ok, in the next refactoring (changing the function args to idx_to_dim_list()) we should consider moving the outer for loop (that calls the  idx_to_dim_list() function) inside the function.